+++ src/polystep/__init__.py
@@ -1,5 +1,5 @@
 """polystep: PyTorch PolyStep Optimizer."""
-__version__ = "0.2.2"
+__version__ = "0.2.4"
 
 __all__ = [
     # Core solver